requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.0.2
django-ckeditor==5.5.0
django-composite-foreignkey==1.0.1
django-crispy-forms==1.7.2
django-import-export==1.0.0
django-js-asset==1.0.0
django-mptt==0.9.0
django-versatileimagefield==1.9
djangorestframework==3.8.2
djangorestframework-recursive==0.1.2
et-xmlfile==1.0.1
gunicorn==19.7.1
idna==2.7
jdcal==1.3
Markdown==2.6.11
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.1
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.12
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
